+++ exchange/test/src/unit/matching/invalid_orders.cpp
@@ -1,5 +1,5 @@
 #include "exchange/config.h"
-#include "exchange/tickers/engine/new_engine.hpp"
+#include "exchange/tickers/engine/engine.hpp"
 #include "test_utils/macros.hpp"
 
 #include <gtest/gtest.h>
